requirements.txt
# Python dependencies
# run: pip install -r requirements.txt
Sphinx
sphinx_rtd_theme
pyaml
